+++ extensions/debugging/src/index.js
@@ -2,6 +2,7 @@ import { getDicomWebClientFromConfig } from './utils';
 import { getCommands } from './commandsModule';
 import { version } from '../package.json';
 import toolbarModule from './toolbarModule';
+import state from './state';
 
 /**
  * Constants
@@ -29,10 +30,14 @@ export default {
    * LIFECYCLE HOOKS
    */
 
-  preRegistration({ appConfig }) {
+  preRegistration({ appConfig, configuration }) {
     const dicomWebClient = getDicomWebClientFromConfig(appConfig);
     if (dicomWebClient) {
       sharedContext.dicomWebClient = dicomWebClient;
+    }
+
+    if (configuration && configuration.mailTo) {
+      state.mailTo = configuration.mailTo;
     }
   },